+++ aeromaps/models/air_transport/aircraft_fleet_and_operations/non_co2/non_co2.py
@@ -14,7 +14,7 @@ class OperationsContrailsSimple(AeroMAPSModel):
         self,
         operations_contrails_final_gain: float,
         operations_contrails_final_overconsumption: float,
-        operations_contrails_start_year: float,
+        operations_contrails_start_year: int,
         operations_contrails_duration: float,
     ) -> Tuple[pd.Series, pd.Series]:
         """Operations contrails gain for ERF calculation."""
